requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
ffmpeg-python>=0.2.0
